forgot to update line 164... (my bad)

+++ src/main/java/io/github/jack1424/realtimeweather/RealTimeWeather.java
@@ -161,7 +161,7 @@ public final class RealTimeWeather extends JavaPlugin {
 			sunsetMinutes += 720;
 
 		LocalTime currentTime = LocalTime.of(cal.get(Calendar.HOUR_OF_DAY), cal.get(Calendar.MINUTE));
-		int currentMinutes = currentTime.getHour() * 60 + currentTime.getMinute();
+		double currentMinutes = currentTime.getHour() * 60 + currentTime.getMinute();
 
 		if (currentMinutes >= sunriseMinutes && currentMinutes < sunsetMinutes) {
 			return (long) (((currentMinutes - sunriseMinutes) / (sunsetMinutes - sunriseMinutes)) * 13569) + 23041;
